♻️ refactor: type 수정
