+++ app/(route)/_components/carousel/NewestEventsCarousel.tsx
@@ -3,7 +3,7 @@
 import { useQuery } from "@tanstack/react-query";
 import VerticalEventCard from "@/components/card/VerticalEventCard";
 import { Api } from "@/api/api";
-import { EventCardType } from "@/types/getBodyType";
+import { Res_Get_Type } from "@/types/getResType";
 import Carousel from "./Carousel";
 
 const NewestEventsCarousel = () => {
@@ -21,7 +21,7 @@ const NewestEvents = () => {
     data: newestEvents,
     isSuccess,
     isLoading,
-  } = useQuery<EventCardType[]>({
+  } = useQuery<Res_Get_Type["eventList"]>({
     queryKey: ["event", "new"],
     queryFn: async () => {
       return instance.get("/event/new");
